# doc/api.yml
openapi: 3.0.3
info:
  title: School Management API
  description: >
    API for managing users, careers, promotions, groups, courses, exams and grades,
    including transcript delivery by email and graduate list exports.
  version: 1.0.0
servers:
  - url: https://ce662nwxk7c5pmpunht2ntmhc40njaoq.lambda-url.eu-west-3.on.aws/
    description: AWS
  - url: http://localhost:9090/
    description: local
security:
  - bearerAuth: []
...
paths:
...
  /promotions:
    get:
      tags: [Promotions]
      summary: List promotions (cohorts)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paged list of promotions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromotionResponsePage'
    post:
      tags: [Promotions]
      summary: Create a promotion
      description: "Requires role: ADMIN."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromotionRequest'
      responses:
        '201':
          description: Promotion created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromotionResponse'
  /promotions/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Promotions]
      summary: Get a promotion by id
      responses:
        '200':
          description: The promotion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromotionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Promotions]
      summary: Update a promotion
      description: "Requires role: ADMIN."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromotionRequest'
      responses:
        '200':
          description: Promotion updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromotionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Promotions]
      summary: Delete a promotion
      description: "Requires role: ADMIN."
      responses:
        '204':
          description: Promotion deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /promotions/{id}/graduates:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Promotions]
      summary: List graduated students of a promotion (JSON)
      description: "Requires role: ADMIN. Feature: promotion results over the 3-year program."
      responses:
        '200':
          description: List of graduated students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
  /promotions/{id}/students-grades:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Promotions]
      summary: Get all students and their grades for a promotion
      description: >
        Requires role: ADMIN. Returns a comprehensive view of every enrolled
        student's grades across all courses and exams in the promotion.
      responses:
        '200':
          description: Students with their grades
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromotionStudentsGradesResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    Role:
      type: string
      enum: [TEACHER, STUDENT, ADMIN]
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ref:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/Role'
...
    PromotionRequest:
      type: object
      required: [label, startYear]
      properties:
        label:
          type: string
        startYear:
          type: integer
    PromotionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
        startYear:
          type: integer
    PromotionResponsePage:
      $ref: '#/components/schemas/PageEnvelope'
...
    PageEnvelope:
      type: object
      properties:
        content:
          type: array
          items: {}
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
...
    PromotionStudentsGradesResponse:
      type: object
      properties:
        promotionId:
          type: string
          format: uuid
        promotionLabel:
          type: string
        startYear:
          type: integer
        students:
          type: array
          items:
            $ref: '#/components/schemas/StudentWithGrades'
    StudentWithGrades:
      type: object
      properties:
        studentId:
          type: string
          format: uuid
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
        courseGrades:
          type: array
          items:
            $ref: '#/components/schemas/PromotionCourseGrade'
    PromotionCourseGrade:
      type: object
      properties:
        courseId:
          type: string
          format: uuid
        courseTitle:
          type: string
        credits:
          type: integer
        examTitle:
          type: string
        grade:
          type: number
          format: double
        examId:
          type: string
          format: uuid